Add files via upload
Ok we got our standardize value. As we saw above the most valuable predictor of this model is gdp_for_year. We can interpreted that the more increase in GDP per year the more suicide rate would increase. Assuming this dataset or sample is a great representation of population.

The second one is suicides_per_100k_pop. 

And the third one it gdp_per_capital.

It's telling us that the more GDP per capital is decreased the more suicide rate will increase.